refactor(generator): extract ContractFinder from SymbolDiscovery
